group-service/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
djangorestframework==3.16.1
kafka-python==2.2.15
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
